Conflicts:
    src/mongo/db/SConscript
    src/mongo/db/auth/SConscript
    src/mongo/db/mongod_main.cpp
    src/mongo/db/storage/SConscript
    src/mongo/db/storage/storage_engine_init.cpp
    src/mongo/db/storage/wiredtiger/wiredtiger_kv_engine.cpp
    src/mongo/db/storage/wiredtiger/wiredtiger_kv_engine.h
    src/mongo/dbtests/SConscript
    src/mongo/s/SConscript
    src/mongo/s/mongos_main.cpp

+++ jstests/core/collation.js
@@ -4,7 +4,7 @@
 // @tags: [
 //   # The test runs commands that are not allowed with security token: applyOps,
 //   # cloneCollectionAsCapped, mapReduce, reIndex.
-//   not_allowed_with_security_token,
+//   not_allowed_with_signed_security_token,
 //   assumes_no_implicit_collection_creation_after_drop,
 //   # Asserts that some queries use a collection scan.
 //   assumes_no_implicit_index_creation,
@@ -21,6 +21,7 @@
 
 // Integration tests for the collation feature.
 import {
+    getOptimizer,
     getPlanStage,
     getSingleNodeExplain,
     getWinningPlan,
@@ -268,15 +269,20 @@ assertIndexHasCollation({e: 1}, {locale: "simple"});
 
 // Test that an index with a non-simple collation contains collator-generated comparison keys
 // rather than the verbatim indexed strings.
-coll = testDb.collation_index2;
-coll.drop();
-assert.commandWorked(coll.createIndex({a: 1}, {collation: {locale: "fr_CA"}}));
-assert.commandWorked(coll.createIndex({b: 1}));
-assert.commandWorked(coll.insert({a: "foo", b: "foo"}));
-assert.eq(1, coll.find().collation({locale: "fr_CA"}).hint({a: 1}).returnKey().itcount());
-assert.neq("foo", coll.find().collation({locale: "fr_CA"}).hint({a: 1}).returnKey().next().a);
-assert.eq(1, coll.find().collation({locale: "fr_CA"}).hint({b: 1}).returnKey().itcount());
-assert.eq("foo", coll.find().collation({locale: "fr_CA"}).hint({b: 1}).returnKey().next().b);
+if (!TestData.isCursorHintsToQuerySettings) {
+    // This guard excludes this test case from being run on the cursor_hints_to_query_settings
+    // suite. The suite replaces cursor hints with query settings. Query settings do not force
+    // indexes, and therefore empty filter will result in collection scans.
+    coll = testDb.collation_index2;
+    coll.drop();
+    assert.commandWorked(coll.createIndex({a: 1}, {collation: {locale: "fr_CA"}}));
+    assert.commandWorked(coll.createIndex({b: 1}));
+    assert.commandWorked(coll.insert({a: "foo", b: "foo"}));
+    assert.eq(1, coll.find().collation({locale: "fr_CA"}).hint({a: 1}).returnKey().itcount());
+    assert.neq("foo", coll.find().collation({locale: "fr_CA"}).hint({a: 1}).returnKey().next().a);
+    assert.eq(1, coll.find().collation({locale: "fr_CA"}).hint({b: 1}).returnKey().itcount());
+    assert.eq("foo", coll.find().collation({locale: "fr_CA"}).hint({b: 1}).returnKey().next().b);
+}
 
 // Test that a query with a string comparison can use an index with a non-simple collation if it
 // has a matching collation.
@@ -669,16 +675,25 @@ assert.commandWorked(coll.dropIndexes());
 
 // Find should return correct results when collation specified and compatible partial index
 // exists.
-assert.commandWorked(coll.createIndex(
-    {str: 1},
-    {partialFilterExpression: {str: {$lte: "FOO"}}, collation: {locale: "en_US", strength: 2}}));
-assert.eq(
-    1, coll.find({str: "foo"}).collation({locale: "en_US", strength: 2}).hint({str: 1}).itcount());
-assert.commandWorked(coll.insert({_id: 3, str: "goo"}));
-assert.eq(
-    0, coll.find({str: "goo"}).collation({locale: "en_US", strength: 2}).hint({str: 1}).itcount());
-assert.commandWorked(coll.remove({_id: 3}));
-assert.commandWorked(coll.dropIndexes());
+if (!TestData.isCursorHintsToQuerySettings) {
+    // This guard excludes this test case from being run on cursor_hints_to_query_settings suite.
+    // The suite replaces cursor hints with query settings. Query settings do not force
+    // partial/sparse indexes with incomplete result sets as described in SERVER-26413, and as such
+    // will yield different results.
+    assert.commandWorked(coll.createIndex({str: 1}, {
+        partialFilterExpression: {str: {$lte: "FOO"}},
+        collation: {locale: "en_US", strength: 2}
+    }));
+    assert.eq(
+        1,
+        coll.find({str: "foo"}).collation({locale: "en_US", strength: 2}).hint({str: 1}).itcount());
+    assert.commandWorked(coll.insert({_id: 3, str: "goo"}));
+    assert.eq(
+        0,
+        coll.find({str: "goo"}).collation({locale: "en_US", strength: 2}).hint({str: 1}).itcount());
+    assert.commandWorked(coll.remove({_id: 3}));
+    assert.commandWorked(coll.dropIndexes());
+}
 
 // Queries that use a index with a non-matching collation should add a sort
 // stage if needed.
@@ -796,7 +811,20 @@ coll.drop();
 assert.commandWorked(testDb.createCollection(coll.getName(), {collation: {locale: "en_US"}}));
 assert.commandWorked(coll.createIndex({a: 1}, {collation: {locale: "en_US"}}));
 var explain = coll.find({a: "foo"}).explain("queryPlanner");
-assert(isIxscan(testDb, getWinningPlan(explain.queryPlanner)));
+
+// Assert the plan is using an index scan.
+switch (getOptimizer(explain)) {
+    case "classic":
+        assert(isIxscan(testDb, getWinningPlan(explain.queryPlanner)));
+        break;
+    case "CQF":
+        // TODO SERVER-77719: Ensure that the decision for using the scan lines up with CQF
+        // optimizer. M2: allow only collscans, M4: check bonsai behavior for index scan.
+        assert(isCollscan(testDb, getWinningPlan(explain.queryPlanner)));
+        break;
+    default:
+        break;
+}
 
 // Find should select compatible index when no collation specified and collection default
 // collation is "simple".
@@ -805,7 +833,20 @@ coll.drop();
 assert.commandWorked(testDb.createCollection(coll.getName(), {collation: {locale: "simple"}}));
 assert.commandWorked(coll.createIndex({a: 1}, {collation: {locale: "simple"}}));
 var explain = coll.find({a: "foo"}).explain("queryPlanner");
-assert(isIxscan(testDb, getWinningPlan(explain.queryPlanner)));
+
+// Assert the plan is using an index scan.
+switch (getOptimizer(explain)) {
+    case "classic":
+        assert(isIxscan(testDb, getWinningPlan(explain.queryPlanner)));
+        break;
+    case "CQF":
+        // TODO SERVER-77719: Ensure that the decision for using the scan lines up with CQF
+        // optimizer. M2: allow only collscans, M4: check bonsai behavior for index scan.
+        assert(isCollscan(testDb, getWinningPlan(explain.queryPlanner)));
+        break;
+    default:
+        break;
+}
 
 // Find should not use index when no collation specified, index collation is "simple", and
 // collection has a non-"simple" default collation.
